cart

+++ E-Commerce.Core/Models/OrderItem.cs
@@ -9,10 +9,10 @@ namespace E_Commerce.Core.Models
     public class OrderItem
     {
         public int Id { get; set; }
-        public int ProductId { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int ProductId  { get; set; }
+        public double UnitPrice   { get; set; }
         public int Quantity { get; set; }
-        public decimal Total { get; set; }
+        public double Total { get; set; }
         public string OrderId { get; set; }
         public Order Order { get; set; }
     }
